@@ -7,21 +7,26 @@
   "repository": "transitive-bullshit/nextjs-notion-starter-kit",
   "license": "MIT",
   "engines": {
-    "node": ">=16"
+    "node": ">=18"
   },
+  "type": "module",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
     "deploy": "vercel deploy",
-    "deps": "run-s deps:*",
-    "deps:update": "[ -z $GITHUB_ACTIONS ] && yarn add notion-client notion-types notion-utils react-notion-x || echo 'Skipping deps:update on CI'",
-    "deps:link": "[ -z $GITHUB_ACTIONS ] && yarn link notion-client notion-types notion-utils react-notion-x || echo 'Skipping deps:link on CI'",
+    "deps:upgrade": "[ -z $GITHUB_ACTIONS ] && pnpm up -L notion-client notion-types notion-utils react-notion-x || echo 'Skipping deps:update on CI'",
+    "deps:link": "[ -z $GITHUB_ACTIONS ] && run-s deps:link:* || echo 'Skipping deps:update on CI'",
+    "deps:unlink": "[ -z $GITHUB_ACTIONS ] && pnpm add notion-client notion-types notion-utils react-notion-x || echo 'Skipping deps:update on CI'",
+    "deps:link:notion-types": "pnpm link ../react-notion-x/packages/notion-types",
+    "deps:link:notion-utils": "pnpm link ../react-notion-x/packages/notion-utils",
+    "deps:link:notion-client": "pnpm link ../react-notion-x/packages/notion-client",
+    "deps:link:react-notion-x": "pnpm link ../react-notion-x/packages/react-notion-x",
     "analyze": "cross-env ANALYZE=true next build",
     "analyze:server": "cross-env BUNDLE_ANALYZE=server next build",
     "analyze:browser": "cross-env BUNDLE_ANALYZE=browser next build",
     "test": "run-p test:*",
-    "test:lint": "eslint '**/*.{ts,tsx}'",
+    "test:lint": "eslint .",
     "test:prettier": "prettier '**/*.{js,jsx,ts,tsx}' --check"
   },
   "dependencies": {
@@ -30,44 +35,37 @@
     "@keyvhq/core": "^1.6.9",
     "@keyvhq/redis": "^1.6.10",
     "@react-icons/all-files": "^4.1.0",
-    "@vercel/og": "^0.0.19",
-    "classnames": "^2.3.1",
-    "date-fns": "^2.28.0",
+    "classnames": "^2.5.1",
     "expiry-map": "^2.0.0",
     "fathom-client": "^3.4.1",
-    "got": "^12.0.3",
-    "isomorphic-unfetch": "^3.1.0",
-    "lqip-modern": "^1.2.0",
-    "next": "^12.3.1",
-    "notion-client": "^6.15.6",
-    "notion-types": "^6.15.6",
-    "notion-utils": "^6.15.6",
-    "p-map": "^5.3.0",
-    "p-memoize": "^6.0.1",
-    "posthog-js": "^1.20.2",
+    "ky": "^1.7.2",
+    "lqip-modern": "^2.2.1",
+    "next": "^15.0.3",
+    "notion-client": "^7.1.3",
+    "notion-types": "^7.1.3",
+    "notion-utils": "^7.1.3",
+    "p-map": "^7.0.2",
+    "p-memoize": "^7.1.1",
+    "posthog-js": "^1.181.0",
+    "prismjs": "^1.29.0",
     "react": "^18.2.0",
     "react-body-classname": "^1.3.1",
     "react-dom": "^18.2.0",
-    "react-notion-x": "^6.15.6",
-    "react-tweet-embed": "^2.0.0",
-    "react-use": "^17.3.2",
+    "react-notion-x": "^7.2.3",
+    "react-tweet": "^3.2.1",
+    "react-use": "^17.4.2",
     "rss": "^1.2.2"
   },
   "devDependencies": {
-    "@next/bundle-analyzer": "^12.3.1",
-    "@trivago/prettier-plugin-sort-imports": "^3.3.1",
-    "@types/node": "^18.8.5",
+    "@fisch0920/eslint-config": "^1.4.0",
+    "@next/bundle-analyzer": "^15.0.2",
+    "@types/node": "^22.8.6",
     "@types/react": "^18.0.21",
-    "@typescript-eslint/eslint-plugin": "^5.40.0",
-    "@typescript-eslint/parser": "^5.40.0",
     "cross-env": "^7.0.2",
-    "eslint": "^8.25.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-react": "^7.31.10",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "npm-run-all": "^4.1.5",
-    "prettier": "^2.7.1",
-    "typescript": "^4.8.4"
+    "eslint": "^8.57.1",
+    "npm-run-all2": "^7.0.1",
+    "prettier": "^3.3.3",
+    "typescript": "^5.6.3"
   },
   "overrides": {
     "cacheable-request": {
