@@ -1,9 +1,9 @@
-import { Block } from 'notion-types'
-import { defaultMapImageUrl } from 'react-notion-x'
+import { type Block } from 'notion-types'
+import { defaultMapImageUrl } from 'notion-utils'
 
 import { defaultPageCover, defaultPageIcon } from './config'
 
-export const mapImageUrl = (url: string, block: Block) => {
+export const mapImageUrl = (url: string | undefined, block: Block) => {
   if (url === defaultPageCover || url === defaultPageIcon) {
     return url
   }
