@@ -1,4 +1,4 @@
-import * as types from './types'
+import type * as types from './types'
 
 export interface SiteConfig {
   rootNotionPageId: string
@@ -16,7 +16,7 @@ export interface SiteConfig {
   newsletter?: string
   youtube?: string
   zhihu?: string
-  mastodon?: string;
+  mastodon?: string
 
   defaultPageIcon?: string | null
   defaultPageCover?: string | null
