@@ -1,9 +1,12 @@
-import { NextApiRequest, NextApiResponse } from 'next'
+import { type NextApiRequest, type NextApiResponse } from 'next'
 
-import * as types from '../../lib/types'
+import type * as types from '../../lib/types'
 import { search } from '../../lib/notion'
 
-export default async (req: NextApiRequest, res: NextApiResponse) => {
+export default async function searchNotion(
+  req: NextApiRequest,
+  res: NextApiResponse
+) {
   if (req.method !== 'POST') {
     return res.status(405).send({ error: 'method not allowed' })
   }
