@@ -1,8 +1,8 @@
 import type { GetServerSideProps } from 'next'
 
+import type { SiteMap } from '@/lib/types'
 import { host } from '@/lib/config'
 import { getSiteMap } from '@/lib/get-site-map'
-import type { SiteMap } from '@/lib/types'
 
 export const getServerSideProps: GetServerSideProps = async ({ req, res }) => {
   if (req.method !== 'GET') {
@@ -54,4 +54,6 @@ const createSitemap = (siteMap: SiteMap) =>
   </urlset>
 `
 
-export default () => null
+export default function noop() {
+  return null
+}
